+++ src/modules/collection/collection-view.test.js
@@ -44,7 +44,6 @@ describe("collection view factory", () => {
       "collectionViewListButton",
       "collectionViewGridButton",
       "collectionViewSwatchButton",
-      "collectionCollapseAllButton",
       "collectionFilterPublishedButton",
       "collectionFilterFavoritesButton",
       "collectionSelectionBar",